moved components inside './views/*' and consolidated them in './components/*
