+++ www/src/components/plugins/package-readme.js
@@ -5,9 +5,9 @@ import PropTypes from "prop-types"
 import { Helmet } from "react-helmet"
 
 import { Link } from "gatsby"
-import Container from "../../components/container"
-import MarkdownPageFooter from "../../components/markdown-page-footer"
-import FooterLinks from "../../components/shared/footer-links"
+import Container from "../container"
+import MarkdownPageFooter from "../markdown-page-footer"
+import FooterLinks from "../shared/footer-links"
 import GithubIcon from "react-icons/lib/go/mark-github"
 import GatsbyIcon from "./gatsby-monogram"
