+++ www/src/components/shared/thumbnail.js
