+++ www/src/components/showcase/showcase-details.js
@@ -8,11 +8,11 @@ import Img from "gatsby-image"
 import qs from "qs"
 
 import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
-import Layout from "../../components/layout"
-import Modal from "../../components/modal"
-import ShareMenu from "../../components/share-menu"
-import Button from "../../components/button"
-import Screenshot from "../../views/shared/screenshot"
+import Layout from "../layout"
+import Modal from "../modal"
+import ShareMenu from "../share-menu"
+import Button from "../button"
+import Screenshot from "../shared/screenshot"
 
 import MdArrowUpward from "react-icons/lib/md/arrow-upward"
 import MdLink from "react-icons/lib/md/link"
