+++ www/src/components/starter-library/starter-list.js
@@ -89,7 +89,11 @@ const StartersList = ({ urlState, starters, count, sortRecent }) => {
                 />
                 <div sx={meta}>
                   <div
-                    css={{ display: `flex`, justifyContent: `space-between`, alignItems: `start` }}
+                    css={{
+                      display: `flex`,
+                      justifyContent: `space-between`,
+                      alignItems: `start`,
+                    }}
                   >
                     <span
                       sx={{
