+++ www/src/components/starter/installation.js
