+++ www/src/pages/showcase.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react"
 import { graphql } from "gatsby"
 
-import ShowcaseView from "../views/showcase"
+import ShowcaseView from "../components/showcase"
 
 class ShowcasePage extends Component {
   render() {
