+++ www/src/templates/template-creator-details.js
@@ -5,8 +5,8 @@ import { graphql, Link } from "gatsby"
 import Layout from "../components/layout"
 import { Helmet } from "react-helmet"
 import Img from "gatsby-image"
-import CreatorsHeader from "../views/creators/creators-header"
-import Badge from "../views/creators/badge"
+import CreatorsHeader from "../components/creators/creators-header"
+import Badge from "../components/creators/badge"
 import FooterLinks from "../components/shared/footer-links"
 import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 import GithubIcon from "react-icons/lib/go/mark-github"
